Add new utility functions
Certain utility functions have been separated from the
`FileTransfer.svelte` view to their own files.
